src/scraper/parser.py
from bs4 import BeautifulSoup
from typing import Dict, List, Optional
import re
from datetime import datetime
from loguru import logger

class ListingParser:
    """Parser for Kleinanzeigen listing pages"""
    
    def __init__(self):
        self.soup = None
        
    def parse_search_results(self, html: str) -> List[Dict]:
        """Parse search results page"""
        # TODO: Implement search results parser
        pass
        
    def parse_listing_details(self, html: str) -> Dict:
        """Parse individual listing details"""
        # TODO: Implement listing details parser
        pass
        
    def extract_contact_info(self, listing_data: Dict) -> Dict:
        """Extract contact information from listing"""
        # TODO: Implement contact info extraction
        pass
        
    def clean_price(self, price_str: str) -> float:
        """Clean and convert price string to float"""
        # TODO: Implement price cleaning
        pass