requirements.txt
# Web scraping
selenium==4.18.1
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
lxml==5.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# HTTP requests
requests==2.31.0
urllib3==2.2.0

# Data processing
pandas==2.2.0
numpy==1.26.3

# Configuration
python-dotenv==1.0.1
pyyaml==6.0.1

# Scheduling
schedule==1.2.0
APScheduler==3.10.4

# Logging
loguru==0.7.2

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development tools
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1